classification/requirements.txt
numpy==1.19.0
pandas==1.0.5
requests==2.24.0
mysqlclient==2.0.1

cython
fasttext
mecab-python3==1.0.1

unidic-lite==1.0.6

Flask==1.1.2
